+++ lib/src/eval/shared/stdlib/core/collection.dart
@@ -5,7 +5,6 @@ import 'package:dart_eval/dart_eval_bridge.dart';
 import 'package:dart_eval/src/eval/shared/stdlib/core/iterator.dart';
 import 'package:dart_eval/stdlib/core.dart';
 import 'package:dart_eval/src/eval/runtime/exception.dart';
-import 'package:dart_eval/src/eval/runtime/override.dart';
 
 part 'iterable.dart';
 part 'list.dart';