+++ lib/src/eval/shared/stdlib/core/iterator.dart
@@ -1,6 +1,5 @@
 import 'package:dart_eval/dart_eval.dart';
 import 'package:dart_eval/dart_eval_bridge.dart';
-import 'package:dart_eval/src/eval/runtime/override.dart';
 import 'package:dart_eval/stdlib/core.dart';
 
 class $Iterator<E> implements Iterator<E>, $Instance {